requirements.txt
google-adk==1.17.0
google-cloud-aiplatform==1.72.0
python-dotenv==1.2.1
pydantic==2.12.4